Style:art:: Add Extension for preview
- UIViewController에 대한 preview를 더 쉽게 사용할 수 있도록 extension 추가
