+++ IntoHistory/HeroDetailViewController.swift
@@ -117,50 +117,6 @@ class HeroDetailViewController: UIViewController {
     }
 }
 
-enum DeviceType {
-    case iPhoneSE2
-    case iPhone8
-    case iPhone12Pro
-    case iPhone12ProMax
-    case iPhone13
-
-    func name() -> String {
-        switch self {
-        case .iPhoneSE2:
-            return "iPhone SE"
-        case .iPhone8:
-            return "iPhone 8"
-        case .iPhone12Pro:
-            return "iPhone 12 Pro"
-        case .iPhone12ProMax:
-            return "iPhone 12 Pro Max"
-        case .iPhone13:
-            return "iPhone 13"
-        }
-    }
-}
-
-#if canImport(SwiftUI) && DEBUG
-import SwiftUI
-extension UIViewController {
-
-    private struct Preview: UIViewControllerRepresentable {
-        let viewController: UIViewController
-
-        func makeUIViewController(context: Context) -> UIViewController {
-            return viewController
-        }
-
-        func updateUIViewController(_ uiViewController: UIViewController, context: Context) {
-        }
-    }
-
-    func showPreview(_ deviceType: DeviceType = .iPhone13) -> some View {
-        Preview(viewController: self).previewDevice(PreviewDevice(rawValue: deviceType.name()))
-    }
-}
-#endif
-
 #if canImport(SwiftUI) && DEBUG
 import SwiftUI
